+++ packages/atlas/src/views/studio/YppDashboard/tabs/index.ts
@@ -1,0 +1,2 @@
+export * from './YppDashboardMainTab'
+export * from './YppDashboardSettingsTab'
